Changed output on DeleteAccountCommand and ChangePasswordCommand

+++ hiper-quiz/src/main/java/commands/DeleteAccountCommand.java
@@ -5,18 +5,26 @@ import exception.EntityNotFoundException;
 import util.LoggedUser;
 
 import java.io.PrintStream;
+import java.util.Scanner;
 
 public class DeleteAccountCommand implements Command{
     PlayerRepository playerRepository;
     PrintStream out;
+    Scanner in;
 
-    public DeleteAccountCommand(PlayerRepository playerRepository, PrintStream out) {
+    public DeleteAccountCommand(PlayerRepository playerRepository, Scanner in, PrintStream out) {
         this.playerRepository = playerRepository;
         this.out = out;
+        this.in = in;
     }
 
     @Override
     public void action() {
+        out.println("Are you sure you want to delete your account? Y/N");
+        String input = in.nextLine();
+        if(input.equals("N")){
+            return;
+        }
         try {
             playerRepository.deleteById(LoggedUser.getLoggedUser().getId());
             out.println("You have successfully deleted your account.");
